refactor(db): inline into the routes bespoke queries
