+++ src/lib/server/database/drizzle.ts
@@ -1,6 +1,4 @@
-import assert, { fail, strictEqual } from 'node:assert/strict';
-
-import { alias, type PgUpdateSetSource } from 'drizzle-orm/pg-core';
+import { alias } from 'drizzle-orm/pg-core';
 import {
   and,
   asc,
@@ -10,23 +8,16 @@ import {
   eq,
   getTableColumns,
   gte,
-  inArray,
   isNotNull,
   isNull,
   lt,
-  lte,
-  ne,
   or,
   sql,
-  sum,
 } from 'drizzle-orm';
-import { array, date, number, object, parse, string, union } from 'valibot';
 import { drizzle } from 'drizzle-orm/node-postgres';
-import { enumerate, izip } from 'itertools';
 
 import { assertOptional, assertSingle } from '$lib/server/assert';
-import { decryptSecret, encryptSecret } from '$lib/crypto';
-import { Logger } from '$lib/server/telemetry/logger';
+import { coerceDate, coerceNullableDate } from '$lib/coerce';
 import { Tracer } from '$lib/server/telemetry/tracer';
 
 import * as schema from './schema';
@@ -35,13 +26,7 @@ import * as schema from './schema';
 export type { schema };
 
 const SERVICE_NAME = 'database';
-const logger = Logger.byName(SERVICE_NAME);
 const tracer = Tracer.byName(SERVICE_NAME);
-
-const StringArray = array(string());
-const StudentRankingLabRemark = array(
-  object({ labId: string(), labName: string(), remark: string() }),
-);
 
 /** Creates a new database instance. */
 export function init(url: string) {
@@ -52,39 +37,9 @@ export type DrizzleDatabase = ReturnType<typeof init>;
 export type DrizzleTransaction = Parameters<Parameters<DrizzleDatabase['transaction']>[0]>[0];
 export type DbConnection = DrizzleDatabase | DrizzleTransaction;
 
-const ParsableDate = union([string(), number(), date()]);
-function coerceDate(value: unknown) {
-  return new Date(parse(ParsableDate, value));
-}
-
-function coerceNullableDate(value: unknown) {
-  return value === null ? null : coerceDate(value);
-}
-
-function coerceNumber(value: unknown) {
-  switch (typeof value) {
-    case 'number':
-      return value;
-    case 'string': {
-      const parsed = Number.parseFloat(value);
-      if (Number.isNaN(parsed)) throw new CoercionError(`expected a numeric string, got: ${value}`);
-      return parsed;
-    }
-    default:
-      throw new CoercionError('expected a number');
-  }
-}
-
 const isRegistrationClosed = lt(schema.draft.registrationClosedAt, sql`now()`)
   .mapWith(Boolean)
   .as('is_registration_closed');
-
-export class CoercionError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'CoercionError';
-  }
-}
 
 export async function insertDummySession(db: DbConnection, userId: string) {
   return await tracer.asyncSpan('insert-dummy-session', async span => {
@@ -95,76 +50,6 @@ export async function insertDummySession(db: DbConnection, userId: string) {
       .returning({ sessionId: schema.session.id })
       .then(assertSingle);
     return sessionId;
-  });
-}
-
-export async function insertValidSession(db: DbConnection, userId: string, expiredAt: Date) {
-  return await tracer.asyncSpan('insert-valid-session', async span => {
-    span.setAttribute('database.user.id', userId);
-    const { sessionId } = await db
-      .insert(schema.session)
-      .values({ userId, expiredAt })
-      .returning({ sessionId: schema.session.id })
-      .then(assertSingle);
-    return sessionId;
-  });
-}
-
-/** Dev-only: Updates the session's userId for user impersonation. */
-export async function updateSessionUserId(db: DbConnection, sessionId: string, userId: string) {
-  return await tracer.asyncSpan('update-session-user-id', async span => {
-    span.setAttributes({ 'database.session.id': sessionId, 'database.user.id': userId });
-    await db.update(schema.session).set({ userId }).where(eq(schema.session.id, sessionId));
-  });
-}
-
-/** Dev-only: Impersonates another user via the session's Id. */
-export async function impersonateUserBySessionId(
-  db: DbConnection,
-  sessionId: string,
-  email: string,
-) {
-  return await tracer.asyncSpan('impersonate-user-by-session-id', async span => {
-    span.setAttributes({ 'database.session.id': sessionId, 'database.user.email': email });
-    const user = await db
-      .update(schema.session)
-      .set({ userId: schema.user.id })
-      .from(schema.user)
-      .where(and(eq(schema.session.id, sessionId), eq(schema.user.email, email)))
-      .returning({ id: schema.user.id })
-      .then(assertOptional);
-
-    if (typeof user === 'undefined') {
-      logger.error('missing session or user');
-      return null;
-    }
-
-    logger.info('Switched to user', { 'user.id': user.id });
-    return user.id;
-  });
-}
-
-export async function getUserById(db: DbConnection, userId: string) {
-  return await tracer.asyncSpan('get-user-by-id', async span => {
-    span.setAttribute('database.user.id', userId);
-    const { id: _, ...columns } = getTableColumns(schema.user);
-    return await db
-      .select(columns)
-      .from(schema.user)
-      .where(eq(schema.user.id, userId))
-      .then(assertSingle);
-  });
-}
-
-export async function getUserFromValidSession(db: DbConnection, sessionId: string) {
-  return await tracer.asyncSpan('get-user-from-valid-session', async span => {
-    span.setAttribute('database.session.id', sessionId);
-    const result = await db.query.session.findFirst({
-      columns: {},
-      with: { user: true },
-      where: ({ id }, { eq }) => eq(id, sessionId),
-    });
-    return result?.user;
   });
 }
 
@@ -209,73 +94,6 @@ export async function upsertOpenIdUser(
   });
 }
 
-export async function updateProfileByUserId(
-  db: DbConnection,
-  userId: string,
-  studentNumber: bigint | null,
-  given: string,
-  family: string,
-) {
-  return await tracer.asyncSpan('update-profile-by-user-id', async span => {
-    span.setAttribute('database.user.id', userId);
-    await db
-      .update(schema.user)
-      .set({
-        studentNumber: sql`coalesce(${schema.user.studentNumber}, ${studentNumber})`,
-        givenName: given,
-        familyName: family,
-      })
-      .where(and(eq(schema.user.id, userId)));
-  });
-}
-
-export async function deleteOpenIdUser(db: DbConnection, userId: string) {
-  return await tracer.asyncSpan('delete-user', async span => {
-    span.setAttribute('database.user.id', userId);
-    await db.delete(schema.user).where(eq(schema.user.id, userId));
-  });
-}
-
-export async function insertNewLab(db: DbConnection, id: string, name: string) {
-  return await tracer.asyncSpan('insert-new-lab', async span => {
-    span.setAttribute('database.lab.id', id);
-    await db.insert(schema.lab).values({ id, name });
-  });
-}
-
-export async function insertNewLabs(db: DbConnection, labs: Pick<schema.Lab, 'id' | 'name'>[]) {
-  return await tracer.asyncSpan('insert-new-labs', async span => {
-    span.setAttribute('database.lab.count', labs.length);
-    await db.insert(schema.lab).values(labs);
-  });
-}
-
-/** Soft-deletion of the lab (a.k.a. "archiving"). */
-export async function deleteLab(db: DbConnection, id: string) {
-  return await tracer.asyncSpan('delete-lab', async span => {
-    span.setAttribute('database.lab.id', id);
-    await db
-      .update(schema.lab)
-      .set({ deletedAt: sql`now()` })
-      .where(eq(schema.lab.id, id));
-  });
-}
-
-/** Hard-deletion of the lab (a.k.a. "dropping"). */
-export async function dropLabs(db: DbConnection, labIds: string[]) {
-  return await tracer.asyncSpan('drop-labs', async span => {
-    span.setAttribute('database.lab.count', labIds.length);
-    await db.delete(schema.lab).where(inArray(schema.lab.id, labIds));
-  });
-}
-
-export async function restoreLab(db: DbConnection, id: string) {
-  return await tracer.asyncSpan('restore-lab', async span => {
-    span.setAttribute('database.lab.id', id);
-    await db.update(schema.lab).set({ deletedAt: null }).where(eq(schema.lab.id, id));
-  });
-}
-
 export async function getLabRegistry(db: DbConnection, activeOnly = true) {
   return await tracer.asyncSpan('get-lab-registry', async span => {
     span.setAttribute('database.lab.active_only', activeOnly);
@@ -288,21 +106,6 @@ export async function getLabRegistry(db: DbConnection, activeOnly = true) {
       })
       .from(targetSchema)
       .orderBy(({ name }) => name);
-  });
-}
-
-export async function getDraftLabRegistry(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-draft-lab-registry', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .select({
-        id: schema.draftLabQuota.labId,
-        name: schema.lab.name,
-      })
-      .from(schema.draftLabQuota)
-      .innerJoin(schema.lab, eq(schema.draftLabQuota.labId, schema.lab.id))
-      .where(eq(schema.draftLabQuota.draftId, draftId))
-      .orderBy(asc(schema.lab.name));
   });
 }
 
@@ -319,48 +122,6 @@ export async function getLabById(db: DbConnection, labId: string) {
   });
 }
 
-export async function getUserNameByEmail(db: DbConnection, email: string) {
-  return await tracer.asyncSpan('get-user-name-by-email', async span => {
-    span.setAttribute('database.user.email', email);
-    return await db
-      .select({ givenName: schema.user.givenName, familyName: schema.user.familyName })
-      .from(schema.user)
-      .where(eq(schema.user.email, email))
-      .then(assertSingle);
-  });
-}
-
-export async function isValidTotalInitialLabQuotaInDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('is-valid-total-initial-lab-quota-in-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const { result } = await db
-      .select({
-        result: sql`bool_or(${schema.draftLabQuota.initialQuota} > 0)`.mapWith(Boolean),
-      })
-      .from(schema.draftLabQuota)
-      .where(eq(schema.draftLabQuota.draftId, draftId))
-      .then(assertSingle);
-    return result;
-  });
-}
-
-export async function getDraftLabQuotaSnapshots(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-draft-lab-quota-snapshots', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .select({
-        labId: schema.draftLabQuota.labId,
-        labName: schema.lab.name,
-        initialQuota: schema.draftLabQuota.initialQuota,
-        lotteryQuota: schema.draftLabQuota.lotteryQuota,
-      })
-      .from(schema.draftLabQuota)
-      .innerJoin(schema.lab, eq(schema.draftLabQuota.labId, schema.lab.id))
-      .where(eq(schema.draftLabQuota.draftId, draftId))
-      .orderBy(asc(schema.lab.name));
-  });
-}
-
 export async function getDraftLabQuotaLabIds(db: DbConnection, draftId: bigint) {
   return await tracer.asyncSpan('get-draft-lab-quota-lab-ids', async span => {
     span.setAttribute('database.draft.id', draftId.toString());
@@ -369,96 +130,6 @@ export async function getDraftLabQuotaLabIds(db: DbConnection, draftId: bigint) 
       .from(schema.draftLabQuota)
       .where(eq(schema.draftLabQuota.draftId, draftId));
     return rows.map(({ labId }) => labId);
-  });
-}
-
-export async function getLabCount(db: DbConnection, activeOnly = true) {
-  return await tracer.asyncSpan('get-lab-count', async span => {
-    span.setAttribute('database.lab.active_only', activeOnly);
-    const targetSchema = activeOnly ? schema.activeLabView : schema.lab;
-    const { result } = await db
-      .select({ result: count(targetSchema.id) })
-      .from(targetSchema)
-      .then(assertSingle);
-    return result;
-  });
-}
-
-export async function getStudentCountInDraft(db: DbConnection, id: bigint) {
-  return await tracer.asyncSpan('get-student-count-in-draft', async span => {
-    span.setAttribute('database.draft.id', id.toString());
-    const { result } = await db
-      .select({ result: count(schema.studentRank.userId) })
-      .from(schema.studentRank)
-      .where(eq(schema.studentRank.draftId, id))
-      .then(assertSingle);
-    return result;
-  });
-}
-
-export async function updateDraftInitialLabQuotas(
-  db: DbConnection,
-  draftId: bigint,
-  quotas: Iterable<readonly [string, number]>,
-) {
-  return await tracer.asyncSpan('update-draft-initial-lab-quotas', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const values = Array.from(
-      quotas,
-      ([labId, initialQuota]): Pick<
-        schema.NewDraftLabQuota,
-        'draftId' | 'labId' | 'initialQuota'
-      > => ({
-        draftId,
-        labId,
-        initialQuota,
-      }),
-    );
-    if (values.length === 0) return;
-
-    await db
-      .insert(schema.draftLabQuota)
-      .values(values)
-      .onConflictDoUpdate({
-        target: [schema.draftLabQuota.draftId, schema.draftLabQuota.labId],
-        set: {
-          initialQuota: sql`excluded.${sql.raw(schema.draftLabQuota.initialQuota.name)}`,
-          updatedAt: sql`now()`,
-        },
-      });
-  });
-}
-
-export async function updateDraftLotteryLabQuotas(
-  db: DbConnection,
-  draftId: bigint,
-  quotas: Iterable<readonly [string, number]>,
-) {
-  return await tracer.asyncSpan('update-draft-lottery-lab-quotas', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const values = Array.from(
-      quotas,
-      ([labId, lotteryQuota]): Pick<
-        schema.NewDraftLabQuota,
-        'draftId' | 'labId' | 'lotteryQuota'
-      > => ({
-        draftId,
-        labId,
-        lotteryQuota,
-      }),
-    );
-    if (values.length === 0) return;
-
-    await db
-      .insert(schema.draftLabQuota)
-      .values(values)
-      .onConflictDoUpdate({
-        target: [schema.draftLabQuota.draftId, schema.draftLabQuota.labId],
-        set: {
-          lotteryQuota: sql`excluded.${sql.raw(schema.draftLabQuota.lotteryQuota.name)}`,
-          updatedAt: sql`now()`,
-        },
-      });
   });
 }
 
@@ -476,171 +147,6 @@ export async function getFacultyAndStaff(db: DbConnection) {
       .from(schema.user)
       .leftJoin(schema.lab, eq(schema.user.labId, schema.lab.id))
       .where(and(eq(schema.user.isAdmin, true), isNotNull(schema.user.googleUserId)));
-  });
-}
-
-export async function getInvitedAdmins(db: DbConnection) {
-  return await tracer.asyncSpan(
-    'get-invited-admins',
-    async () =>
-      await db
-        .select({
-          id: schema.user.id,
-          googleUserId: schema.user.googleUserId,
-          email: schema.user.email,
-          givenName: schema.user.givenName,
-          familyName: schema.user.familyName,
-          avatarUrl: schema.user.avatarUrl,
-          labId: schema.lab.id,
-          labName: schema.lab.name,
-        })
-        .from(schema.user)
-        .leftJoin(schema.lab, eq(schema.user.labId, schema.lab.id))
-        .where(
-          and(
-            eq(schema.user.isAdmin, true),
-            isNull(schema.user.googleUserId),
-            isNull(schema.user.labId),
-          ),
-        ),
-  );
-}
-
-export async function getInvitedHeads(db: DbConnection) {
-  return await tracer.asyncSpan(
-    'get-invited-heads',
-    async () =>
-      await db
-        .select({
-          id: schema.user.id,
-          googleUserId: schema.user.googleUserId,
-          email: schema.user.email,
-          givenName: schema.user.givenName,
-          familyName: schema.user.familyName,
-          avatarUrl: schema.user.avatarUrl,
-          labId: schema.lab.id,
-          labName: schema.lab.name,
-        })
-        .from(schema.user)
-        .leftJoin(schema.lab, eq(schema.user.labId, schema.lab.id))
-        .where(
-          and(
-            eq(schema.user.isAdmin, true),
-            isNull(schema.user.googleUserId),
-            isNotNull(schema.user.labId),
-          ),
-        ),
-  );
-}
-
-export async function getValidStaffEmails(db: DbConnection) {
-  return await tracer.asyncSpan('get-valid-staff-emails', async () => {
-    const results = await db
-      .select({ email: schema.user.email })
-      .from(schema.user)
-      .where(
-        and(
-          eq(schema.user.isAdmin, true),
-          isNull(schema.user.labId),
-          isNotNull(schema.user.googleUserId),
-        ),
-      );
-    return results.map(({ email }) => email);
-  });
-}
-
-/** Ideally invoked from within a transaction. */
-export async function getLabMembers(db: DbConnection, labId: string, draftId?: bigint) {
-  return await tracer.asyncSpan('get-lab-members', async span => {
-    span.setAttribute('database.lab.id', labId);
-    if (typeof draftId !== 'undefined') span.setAttribute('database.draft.id', draftId.toString());
-    const labInfo = await db.query.lab.findFirst({
-      columns: { name: true },
-      where: ({ id }) => eq(id, labId),
-    });
-
-    const heads = await db.query.user.findMany({
-      columns: {
-        email: true,
-        givenName: true,
-        familyName: true,
-        avatarUrl: true,
-      },
-      where: and(
-        isNotNull(schema.user.id),
-        eq(schema.user.labId, labId),
-        eq(schema.user.isAdmin, true),
-      ),
-      orderBy: ({ familyName }) => familyName,
-    });
-
-    const faculty = await db
-      .select({
-        email: schema.user.email,
-        givenName: schema.user.givenName,
-        familyName: schema.user.familyName,
-        avatarUrl: schema.user.avatarUrl,
-      })
-      .from(schema.user)
-      .leftJoin(
-        schema.facultyChoiceUser,
-        eq(schema.user.id, schema.facultyChoiceUser.studentUserId),
-      )
-      .where(
-        and(
-          eq(schema.user.labId, labId),
-          eq(schema.user.isAdmin, false),
-          isNull(schema.facultyChoiceUser.studentUserId),
-        ),
-      );
-
-    const membersQuery =
-      typeof draftId === 'undefined'
-        ? db
-            .select({
-              draftId: schema.labMemberView.draftId,
-              email: schema.labMemberView.email,
-              givenName: schema.labMemberView.givenName,
-              familyName: schema.labMemberView.familyName,
-              avatarUrl: schema.labMemberView.avatarUrl,
-            })
-            .from(schema.labMemberView)
-            .where(eq(schema.labMemberView.draftLab, labId))
-            .orderBy(asc(schema.labMemberView.draftId), asc(schema.labMemberView.familyName))
-        : db
-            .select({
-              draftId: schema.labMemberView.draftId,
-              email: schema.labMemberView.email,
-              givenName: schema.labMemberView.givenName,
-              familyName: schema.labMemberView.familyName,
-              avatarUrl: schema.labMemberView.avatarUrl,
-            })
-            .from(schema.labMemberView)
-            .where(
-              and(
-                eq(schema.labMemberView.draftLab, labId),
-                eq(schema.labMemberView.draftId, draftId),
-              ),
-            )
-            .orderBy(asc(schema.labMemberView.draftId), asc(schema.labMemberView.familyName));
-    return { lab: labInfo?.name, heads, members: await membersQuery, faculty };
-  });
-}
-
-/**
- * Get the latest draft id that a user participated in to be assigned to a certain lab
- * @param userId The id of the user whose draft id is to be identified
- * @param labId The id of the lab they were assigned for
- */
-export async function getUserLabAssignmentDraftId(db: DbConnection, userId: string, labId: string) {
-  return await tracer.asyncSpan('get-user-lab-assignment-draft-id', async span => {
-    span.setAttributes({ 'database.user.id': userId, 'database.lab.id': labId });
-    return await db
-      .select({ draftId: schema.labMemberView.draftId })
-      .from(schema.labMemberView)
-      .where(and(eq(schema.labMemberView.userId, userId), eq(schema.labMemberView.draftLab, labId)))
-      .orderBy(desc(schema.labMemberView.draftId))
-      .then(assertOptional);
   });
 }
 
@@ -713,10 +219,8 @@ export async function getActiveDraft(db: DbConnection) {
         currRound: schema.draft.currRound,
         maxRounds: schema.draft.maxRounds,
         registrationClosedAt: schema.draft.registrationClosedAt,
-        startedAt: schema.draft.startedAt,
         isRegistrationClosed,
         activePeriodStart: sql`lower(${schema.draft.activePeriod})`.mapWith(coerceDate),
-        activePeriodEnd: sql`upper(${schema.draft.activePeriod})`.mapWith(coerceNullableDate),
       })
       .from(schema.draft)
       .where(sql`upper_inf(${schema.draft.activePeriod})`)
@@ -724,314 +228,6 @@ export async function getActiveDraft(db: DbConnection) {
   });
 }
 
-export async function lockLabCatalogForMutation(db: DrizzleTransaction) {
-  return await tracer.asyncSpan(
-    'lock-lab-catalog-for-mutation',
-    async () => await db.execute(sql`LOCK TABLE ${schema.lab} IN ROW EXCLUSIVE MODE`),
-  );
-}
-
-export async function getDraftByIdForShare(db: DrizzleTransaction, id: bigint) {
-  return await tracer.asyncSpan('get-draft-by-id-for-share', async span => {
-    span.setAttribute('database.draft.id', id.toString());
-    return await db
-      .select({
-        currRound: schema.draft.currRound,
-        maxRounds: schema.draft.maxRounds,
-        registrationClosedAt: schema.draft.registrationClosedAt,
-        isRegistrationClosed,
-        activePeriodStart: sql`lower(${schema.draft.activePeriod})`.mapWith(coerceDate),
-        activePeriodEnd: sql`upper(${schema.draft.activePeriod})`.mapWith(coerceNullableDate),
-      })
-      .from(schema.draft)
-      .where(eq(schema.draft.id, id))
-      .for('share')
-      .then(assertOptional);
-  });
-}
-
-export async function hasActiveDraft(db: DbConnection) {
-  return await tracer.asyncSpan('has-active-draft', async () => {
-    const result = await db
-      .select({ one: sql.raw('1') })
-      .from(schema.draft)
-      .where(isNull(sql`upper(${schema.draft.activePeriod})`))
-      .limit(1)
-      .then(assertOptional);
-    return typeof result !== 'undefined';
-  });
-}
-
-export async function getMaxRoundInDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-max-round-in-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const result = await db
-      .select({ maxRounds: schema.draft.maxRounds })
-      .from(schema.draft)
-      .where(eq(schema.draft.id, draftId))
-      .then(assertOptional);
-    return result?.maxRounds;
-  });
-}
-
-export async function getStudentsInDraftTaggedByLab(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-students-in-draft-tagged-by-lab', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .select({
-        id: schema.user.id,
-        email: schema.user.email,
-        givenName: schema.user.givenName,
-        familyName: schema.user.familyName,
-        avatarUrl: schema.user.avatarUrl,
-        studentNumber: schema.user.studentNumber,
-        labs: sql`coalesce(array_agg(${schema.studentRankLab.labId} order by ${schema.studentRankLab.index}) filter (where ${isNotNull(schema.studentRankLab.labId)}), '{}')`
-          .mapWith(value => parse(StringArray, value))
-          .as('labs'),
-        labId: schema.facultyChoiceUser.labId,
-      })
-      .from(schema.studentRank)
-      .innerJoin(schema.user, eq(schema.studentRank.userId, schema.user.id))
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.studentRank.draftId, schema.facultyChoiceUser.draftId),
-          eq(schema.studentRank.userId, schema.facultyChoiceUser.studentUserId),
-        ),
-      )
-      .leftJoin(
-        schema.studentRankLab,
-        and(
-          eq(schema.studentRank.draftId, schema.studentRankLab.draftId),
-          eq(schema.studentRank.userId, schema.studentRankLab.userId),
-        ),
-      )
-      .where(eq(schema.studentRank.draftId, draftId))
-      .groupBy(schema.user.id, schema.facultyChoiceUser.labId)
-      .orderBy(schema.user.familyName);
-  });
-}
-
-/**
- * Typically invoked from within a transaction.
- *
- * For a lab to be auto-acknowledged:
- * 1. The lab has exhausted their entire quota (i.e., no remaining quota).
- * 2. The lab must not be preferred by anyone in the current draft round.
- */
-export async function getLabAndRemainingStudentsInDraftWithLabPreference(
-  db: DbConnection,
-  draftId: bigint,
-  labId: string,
-) {
-  return await tracer.asyncSpan(
-    'get-lab-and-remaining-students-in-draft-with-lab-preference',
-    async span => {
-      span.setAttribute('database.draft.id', draftId.toString());
-      span.setAttribute('database.lab.id', labId);
-      const labInfo = await db
-        .select({
-          name: schema.lab.name,
-          quota: schema.draftLabQuota.initialQuota,
-        })
-        .from(schema.draftLabQuota)
-        .innerJoin(schema.lab, eq(schema.draftLabQuota.labId, schema.lab.id))
-        .where(
-          and(eq(schema.draftLabQuota.draftId, draftId), eq(schema.draftLabQuota.labId, labId)),
-        )
-        .then(assertOptional);
-      if (typeof labInfo === 'undefined') return;
-
-      const students = await db
-        .select({
-          id: schema.user.id,
-          email: schema.user.email,
-          givenName: schema.user.givenName,
-          familyName: schema.user.familyName,
-          avatarUrl: schema.user.avatarUrl,
-          studentNumber: schema.user.studentNumber,
-          remark: schema.studentRankLab.remark,
-        })
-        .from(schema.studentRank)
-        .innerJoin(schema.draft, eq(schema.studentRank.draftId, schema.draft.id))
-        .leftJoin(
-          schema.facultyChoiceUser,
-          and(
-            eq(schema.studentRank.draftId, schema.facultyChoiceUser.draftId),
-            eq(schema.studentRank.userId, schema.facultyChoiceUser.studentUserId),
-          ),
-        )
-        .innerJoin(
-          schema.studentRankLab,
-          and(
-            eq(schema.studentRank.draftId, schema.studentRankLab.draftId),
-            eq(schema.studentRank.userId, schema.studentRankLab.userId),
-          ),
-        )
-        .innerJoin(schema.user, eq(schema.studentRank.userId, schema.user.id))
-        .where(
-          and(
-            eq(schema.studentRank.draftId, draftId),
-            or(
-              isNull(schema.facultyChoiceUser.studentUserId),
-              and(
-                eq(schema.facultyChoiceUser.labId, labId),
-                eq(schema.facultyChoiceUser.round, schema.draft.currRound),
-              ),
-            ),
-            eq(schema.studentRankLab.index, schema.draft.currRound),
-            eq(schema.studentRankLab.labId, labId),
-          ),
-        )
-        .orderBy(schema.user.familyName);
-
-      const researchers = await db
-        .select({
-          id: schema.user.id,
-          round: sql`${schema.facultyChoiceUser.round}`.mapWith(coerceNumber),
-          email: schema.user.email,
-          givenName: schema.user.givenName,
-          familyName: schema.user.familyName,
-          avatarUrl: schema.user.avatarUrl,
-          studentNumber: schema.user.studentNumber,
-        })
-        .from(schema.facultyChoiceUser)
-        .innerJoin(schema.user, eq(schema.facultyChoiceUser.studentUserId, schema.user.id))
-        .where(
-          and(
-            eq(schema.facultyChoiceUser.draftId, draftId),
-            eq(schema.facultyChoiceUser.labId, labId),
-            isNotNull(schema.facultyChoiceUser.round),
-          ),
-        );
-
-      const choice = await db
-        .select({ userId: schema.facultyChoice.userId })
-        .from(schema.facultyChoice)
-        .innerJoin(
-          schema.draft,
-          and(
-            eq(schema.facultyChoice.draftId, schema.draft.id),
-            eq(schema.facultyChoice.round, schema.draft.currRound),
-          ),
-        )
-        .where(
-          and(eq(schema.facultyChoice.draftId, draftId), eq(schema.facultyChoice.labId, labId)),
-        )
-        .then(assertOptional);
-
-      // eslint-disable-next-line @typescript-eslint/init-declarations
-      let submissionSource: 'faculty' | 'system' | undefined;
-      if (typeof choice !== 'undefined')
-        submissionSource = choice.userId === null ? 'system' : 'faculty';
-
-      const remainingQuota = labInfo.quota - researchers.length;
-
-      // eslint-disable-next-line @typescript-eslint/init-declarations
-      let autoAcknowledgeReason: 'quota-exhausted' | 'no-preferences' | undefined;
-      if (remainingQuota <= 0) autoAcknowledgeReason = 'quota-exhausted';
-      else if (students.length === 0) autoAcknowledgeReason = 'no-preferences';
-
-      return {
-        lab: labInfo,
-        students,
-        researchers,
-        submissionSource,
-        remainingQuota,
-        autoAcknowledgeReason,
-      };
-    },
-  );
-}
-
-export async function getLabAutoAcknowledgeStatusInDraftRound(
-  db: DbConnection,
-  draftId: bigint,
-  labId: string,
-  currRound: number,
-) {
-  return await tracer.asyncSpan('get-lab-auto-acknowledge-status-in-draft-round', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.lab.id', labId);
-    span.setAttribute('database.round', currRound);
-
-    const row = await db
-      .select({
-        initialQuota: schema.draftLabQuota.initialQuota,
-        totalDrafted: count(schema.facultyChoiceUser.studentUserId),
-        submissionLabId: schema.facultyChoice.labId,
-        submissionUserId: schema.facultyChoice.userId,
-      })
-      .from(schema.draftLabQuota)
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.facultyChoiceUser.draftId, draftId),
-          eq(schema.facultyChoiceUser.labId, labId),
-        ),
-      )
-      .leftJoin(
-        schema.facultyChoice,
-        and(
-          eq(schema.facultyChoice.draftId, draftId),
-          eq(schema.facultyChoice.round, currRound),
-          eq(schema.facultyChoice.labId, labId),
-        ),
-      )
-      .where(and(eq(schema.draftLabQuota.draftId, draftId), eq(schema.draftLabQuota.labId, labId)))
-      .groupBy(
-        schema.draftLabQuota.initialQuota,
-        schema.facultyChoice.labId,
-        schema.facultyChoice.userId,
-      )
-      .then(assertOptional);
-    if (typeof row === 'undefined') return;
-
-    const { initialQuota, totalDrafted, submissionLabId, submissionUserId } = row;
-
-    // eslint-disable-next-line @typescript-eslint/init-declarations
-    let autoAcknowledgeReason: 'quota-exhausted' | 'no-preferences' | undefined;
-    if (totalDrafted >= initialQuota) {
-      autoAcknowledgeReason = 'quota-exhausted';
-    } else {
-      const { preferrerCount } = await db
-        .select({ preferrerCount: countDistinct(schema.studentRankLab.userId) })
-        .from(schema.studentRankLab)
-        .leftJoin(
-          schema.facultyChoiceUser,
-          and(
-            eq(schema.studentRankLab.draftId, schema.facultyChoiceUser.draftId),
-            eq(schema.studentRankLab.userId, schema.facultyChoiceUser.studentUserId),
-          ),
-        )
-        .where(
-          and(
-            eq(schema.studentRankLab.draftId, draftId),
-            eq(schema.studentRankLab.labId, labId),
-            eq(schema.studentRankLab.index, BigInt(currRound)),
-            or(
-              isNull(schema.facultyChoiceUser.studentUserId),
-              and(
-                eq(schema.facultyChoiceUser.labId, labId),
-                eq(schema.facultyChoiceUser.round, currRound),
-              ),
-            ),
-          ),
-        )
-        .then(assertSingle);
-      if (preferrerCount === 0) autoAcknowledgeReason = 'no-preferences';
-    }
-
-    // eslint-disable-next-line @typescript-eslint/init-declarations
-    let submissionSource: 'faculty' | 'system' | undefined;
-    if (submissionLabId !== null)
-      submissionSource = submissionUserId === null ? 'system' : 'faculty';
-
-    return { autoAcknowledgeReason, submissionSource };
-  });
-}
-
-/** Typically invoked from within a transaction. */
 export async function autoAcknowledgeLabsWithoutPreferences(
   db: DrizzleTransaction,
   draftId: bigint,
@@ -1128,182 +324,6 @@ export async function autoAcknowledgeLabsWithoutPreferences(
   });
 }
 
-export async function getLabQuotaAndSelectedStudentCountInDraft(
-  db: DbConnection,
-  draftId: bigint,
-  labId: string,
-) {
-  return await tracer.asyncSpan('get-lab-quota-and-selected-student-count-in-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.lab.id', labId);
-    const labInfo = await db
-      .select({ quota: schema.draftLabQuota.initialQuota })
-      .from(schema.draftLabQuota)
-      .where(and(eq(schema.draftLabQuota.draftId, draftId), eq(schema.draftLabQuota.labId, labId)))
-      .then(assertOptional);
-
-    const draftCount = await db
-      .select({ studentCount: count(schema.facultyChoiceUser.studentUserId) })
-      .from(schema.facultyChoiceUser)
-      .where(
-        and(
-          eq(schema.facultyChoiceUser.draftId, draftId),
-          eq(schema.facultyChoiceUser.labId, labId),
-        ),
-      )
-      .then(assertSingle);
-
-    return {
-      quota: labInfo?.quota,
-      selected: draftCount.studentCount,
-    };
-  });
-}
-
-export async function getFacultyChoiceForLabInDraftRound(
-  db: DbConnection,
-  draftId: bigint,
-  round: number,
-  labId: string,
-) {
-  return await tracer.asyncSpan('get-faculty-choice-for-lab-in-draft-round', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.round', round);
-    span.setAttribute('database.lab.id', labId);
-    return await db
-      .select({
-        draftId: schema.facultyChoice.draftId,
-        round: schema.facultyChoice.round,
-        labId: schema.facultyChoice.labId,
-        userId: schema.facultyChoice.userId,
-      })
-      .from(schema.facultyChoice)
-      .where(
-        and(
-          eq(schema.facultyChoice.draftId, draftId),
-          eq(schema.facultyChoice.round, round),
-          eq(schema.facultyChoice.labId, labId),
-        ),
-      )
-      .then(assertOptional);
-  });
-}
-
-export async function getLabSelectedStudentCountInDraftRound(
-  db: DbConnection,
-  draftId: bigint,
-  labId: string,
-  round: number,
-) {
-  return await tracer.asyncSpan('get-lab-selected-student-count-in-draft-round', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.lab.id', labId);
-    span.setAttribute('database.round', round);
-    const { studentCount } = await db
-      .select({ studentCount: count(schema.facultyChoiceUser.studentUserId) })
-      .from(schema.facultyChoiceUser)
-      .where(
-        and(
-          eq(schema.facultyChoiceUser.draftId, draftId),
-          eq(schema.facultyChoiceUser.labId, labId),
-          eq(schema.facultyChoiceUser.round, round),
-        ),
-      )
-      .then(assertSingle);
-    return studentCount;
-  });
-}
-
-/**
- * Validates that all submitted students chose the specified lab for the given round.
- * Returns a Set of valid student user IDs that match the criteria.
- */
-export async function validateStudentsChoseLabInRound(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  labId: string,
-  round: number,
-  studentUserIds: string[],
-) {
-  return await tracer.asyncSpan('validate-students-chose-lab-in-round', async span => {
-    span.setAttributes({
-      'database.draft.id': draftId.toString(),
-      'database.lab.id': labId,
-      'database.round': round,
-      'database.student.count': studentUserIds.length,
-    });
-
-    if (studentUserIds.length === 0) return new Set();
-
-    // Subquery to find students already drafted by a DIFFERENT lab (or different round of same lab).
-    // These students are invalid selections regardless of their preference.
-    const draftedByOtherLab = db
-      .select({ studentUserId: schema.facultyChoiceUser.studentUserId })
-      .from(schema.facultyChoiceUser)
-      .where(
-        and(
-          eq(schema.facultyChoiceUser.draftId, draftId),
-          or(
-            ne(schema.facultyChoiceUser.labId, labId),
-            ne(schema.facultyChoiceUser.round, round),
-            isNull(schema.facultyChoiceUser.round),
-          ),
-        ),
-      );
-
-    const validRows = await db
-      .select({ userId: schema.studentRankLab.userId })
-      .from(schema.studentRankLab)
-      .where(
-        and(
-          eq(schema.studentRankLab.draftId, draftId),
-          eq(schema.studentRankLab.labId, labId),
-          eq(schema.studentRankLab.index, BigInt(round)),
-          inArray(schema.studentRankLab.userId, studentUserIds),
-          sql`${schema.studentRankLab.userId} not in (${draftedByOtherLab})`,
-        ),
-      )
-      .for('update');
-
-    return new Set(validRows.map(({ userId }) => userId));
-  });
-}
-
-export async function initDraft(
-  db: DrizzleTransaction,
-  maxRounds: number,
-  registrationClosedAt: Date,
-) {
-  return await tracer.asyncSpan('init-draft', async span => {
-    span.setAttribute('database.draft.max_rounds', maxRounds);
-
-    // Blocks further modifications to the lab catalog until the end of the transaction.
-    await db.execute(sql`LOCK TABLE ${schema.lab} IN SHARE MODE`);
-
-    const draft = await db
-      .insert(schema.draft)
-      .values({ maxRounds, registrationClosedAt })
-      .returning({
-        id: schema.draft.id,
-        activePeriodStart: sql`lower(${schema.draft.activePeriod})`.mapWith(coerceDate),
-      })
-      .then(assertSingle);
-
-    const labs = await db.select({ labId: schema.activeLabView.id }).from(schema.activeLabView);
-    if (labs.length > 0)
-      await db.insert(schema.draftLabQuota).values(
-        labs.map(
-          ({ labId }): Pick<schema.NewDraftLabQuota, 'draftId' | 'labId'> => ({
-            draftId: draft.id,
-            labId,
-          }),
-        ),
-      );
-
-    return draft;
-  });
-}
-
 export async function incrementDraftRound(db: DbConnection, draftId: bigint) {
   return await tracer.asyncSpan('increment-draft-round', async span => {
     span.setAttribute('database.draft.id', draftId.toString());
@@ -1318,484 +338,6 @@ export async function incrementDraftRound(db: DbConnection, draftId: bigint) {
         maxRounds: schema.draft.maxRounds,
       })
       .then(assertOptional);
-  });
-}
-
-export async function markDraftAsStarted(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('mark-draft-as-started', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .update(schema.draft)
-      .set({ startedAt: sql`now()` })
-      .where(eq(schema.draft.id, draftId))
-      .returning({ startedAt: sql`${schema.draft.startedAt}`.mapWith(coerceDate) })
-      .then(assertSingle);
-  });
-}
-
-export async function randomizeRemainingStudents(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('randomize-remaining-students', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const results = await db
-      .select({ userId: schema.studentRank.userId })
-      .from(schema.studentRank)
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.studentRank.draftId, schema.facultyChoiceUser.draftId),
-          eq(schema.studentRank.userId, schema.facultyChoiceUser.studentUserId),
-        ),
-      )
-      .where(
-        and(
-          eq(schema.studentRank.draftId, draftId),
-          isNull(schema.facultyChoiceUser.studentUserId),
-        ),
-      )
-      .orderBy(sql`random()`);
-    return results.map(({ userId }) => userId);
-  });
-}
-
-export async function concludeDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('conclude-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const { activePeriodEnd } = await db
-      .update(schema.draft)
-      .set({
-        activePeriod: sql`tstzrange(lower(${schema.draft.activePeriod}), coalesce(upper(${schema.draft.activePeriod}), now()), '[)')`,
-      })
-      .where(eq(schema.draft.id, draftId))
-      .returning({ activePeriodEnd: sql`upper(${schema.draft.activePeriod})`.mapWith(coerceDate) })
-      .then(assertSingle);
-    return activePeriodEnd;
-  });
-}
-
-export async function beginDraftReview(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('begin-draft-review', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .update(schema.draft)
-      .set({ currRound: null })
-      .where(eq(schema.draft.id, draftId))
-      .returning({ maxRounds: schema.draft.maxRounds })
-      .then(assertOptional);
-  });
-}
-
-export async function insertStudentRanking(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  userId: string,
-  labs: string[],
-  remarks: string[],
-) {
-  return await tracer.asyncSpan('insert-student-ranking', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.user.id', userId);
-    await db
-      .insert(schema.studentRank)
-      .values({ draftId, userId })
-      .onConflictDoNothing({ target: [schema.studentRank.draftId, schema.studentRank.userId] });
-
-    for (const [index, [labId, remark]] of enumerate(izip(labs, remarks)))
-      await db
-        .insert(schema.studentRankLab)
-        .values({ draftId, userId, labId, index: BigInt(index + 1), remark });
-  });
-}
-
-export async function getStudentRankings(db: DbConnection, draftId: bigint, userId: string) {
-  return await tracer.asyncSpan('get-student-rankings', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.user.id', userId);
-    const sub = db
-      .select({
-        createdAt: schema.studentRank.createdAt,
-        index: schema.studentRankLab.index,
-        labId: schema.studentRankLab.labId,
-        remark: schema.studentRankLab.remark,
-      })
-      .from(schema.studentRank)
-      .leftJoin(
-        schema.studentRankLab,
-        and(
-          eq(schema.studentRank.draftId, schema.studentRankLab.draftId),
-          eq(schema.studentRank.userId, schema.studentRankLab.userId),
-        ),
-      )
-      .where(and(eq(schema.studentRank.draftId, draftId), eq(schema.studentRank.userId, userId)))
-      .as('_');
-
-    return await db
-      .select({
-        createdAt: sub.createdAt,
-        labRemarks:
-          sql`coalesce(jsonb_agg(jsonb_build_object('labId', ${sub.labId}, 'labName', ${schema.lab.name}, 'remark', ${sub.remark}) ORDER BY ${sub.index}) filter (where ${isNotNull(sub.labId)}), '[]'::jsonb)`.mapWith(
-            vals => parse(StudentRankingLabRemark, vals),
-          ),
-      })
-      .from(sub)
-      .leftJoin(schema.lab, eq(sub.labId, schema.lab.id))
-      .groupBy(sub.createdAt)
-      .then(assertOptional);
-  });
-}
-
-export async function getCandidateSenders(db: DbConnection) {
-  return await tracer.asyncSpan('get-candidate-senders', async () => {
-    return await db
-      .select({
-        id: schema.user.id,
-        email: schema.user.email,
-        givenName: schema.user.givenName,
-        familyName: schema.user.familyName,
-        avatarUrl: schema.user.avatarUrl,
-        isActive: isNotNull(schema.designatedSender.candidateSenderUserId).mapWith(Boolean),
-      })
-      .from(schema.candidateSender)
-      .innerJoin(schema.user, eq(schema.candidateSender.userId, schema.user.id))
-      .leftJoin(
-        schema.designatedSender,
-        eq(schema.candidateSender.userId, schema.designatedSender.candidateSenderUserId),
-      )
-      .where(
-        and(isNotNull(schema.user.id), eq(schema.user.isAdmin, true), isNull(schema.user.labId)),
-      )
-      .orderBy(schema.user.familyName);
-  });
-}
-
-/**
- * A designated sender is an admin (i.e., `is_admin=True` and `lab_id=NULL`) with valid
- * OAuth 2.0 credentials with an access token that has not yet expired.
- */
-export async function getDesignatedSenderCredentialsForUpdate(
-  db: DrizzleTransaction,
-  encryptionKey: CryptoKey,
-) {
-  return await tracer.asyncSpan('get-designated-sender-credentials', async () => {
-    const sender = await db
-      .select({
-        id: schema.user.id,
-        email: schema.user.email,
-        givenName: schema.user.givenName,
-        familyName: schema.user.familyName,
-        accessTokenIv: schema.candidateSender.accessTokenIv,
-        accessTokenCipher: schema.candidateSender.accessTokenCipher,
-        refreshTokenIv: schema.candidateSender.refreshTokenIv,
-        refreshTokenCipher: schema.candidateSender.refreshTokenCipher,
-        scopes: schema.candidateSender.scopes,
-        isValid: gte(schema.candidateSender.expiredAt, sql`now()`).mapWith(Boolean),
-      })
-      .from(schema.designatedSender)
-      .innerJoin(
-        schema.candidateSender,
-        eq(schema.designatedSender.candidateSenderUserId, schema.candidateSender.userId),
-      )
-      .innerJoin(schema.user, eq(schema.designatedSender.candidateSenderUserId, schema.user.id))
-      .where(
-        and(
-          isNotNull(schema.user.googleUserId),
-          eq(schema.user.isAdmin, true),
-          isNull(schema.user.labId),
-        ),
-      )
-      .for('update')
-      .then(assertOptional);
-
-    if (typeof sender === 'undefined') return;
-
-    const [accessToken, refreshToken] = await Promise.all([
-      decryptSecret(encryptionKey, sender.accessTokenIv, sender.accessTokenCipher),
-      decryptSecret(encryptionKey, sender.refreshTokenIv, sender.refreshTokenCipher),
-    ]);
-
-    return {
-      id: sender.id,
-      email: sender.email,
-      givenName: sender.givenName,
-      familyName: sender.familyName,
-      accessToken,
-      refreshToken,
-      scopes: sender.scopes,
-      isValid: sender.isValid,
-    };
-  });
-}
-
-export async function updateCandidateSender(
-  db: DbConnection,
-  userId: string,
-  expiresIn: number,
-  scopes: string[],
-  encryptionKey: CryptoKey,
-  accessToken: string,
-  refreshToken?: string | undefined,
-) {
-  return await tracer.asyncSpan('update-candidate-sender', async span => {
-    span.setAttributes({
-      'database.user.id': userId,
-      'database.candidate_sender.expires_in': expiresIn,
-    });
-
-    const encryptedAccessToken = await encryptSecret(encryptionKey, accessToken);
-    const update: PgUpdateSetSource<typeof schema.candidateSender> = {
-      userId,
-      scopes,
-      accessTokenIv: Buffer.from(encryptedAccessToken.iv),
-      accessTokenCipher: Buffer.from(encryptedAccessToken.cipher),
-      refreshTokenIv: schema.candidateSender.refreshTokenIv,
-      refreshTokenCipher: schema.candidateSender.refreshTokenCipher,
-      expiredAt: sql`now() + make_interval(secs => ${expiresIn})`,
-    };
-
-    if (typeof refreshToken !== 'undefined') {
-      const { iv, cipher } = await encryptSecret(encryptionKey, refreshToken);
-      update.refreshTokenIv = Buffer.from(iv);
-      update.refreshTokenCipher = Buffer.from(cipher);
-    }
-
-    const { rowCount } = await db
-      .update(schema.candidateSender)
-      .set(update)
-      .where(eq(schema.candidateSender.userId, userId));
-    logger.debug('updated candidate sender', { rowCount });
-  });
-}
-
-export async function upsertCandidateSender(
-  db: DbConnection,
-  userId: string,
-  expiredAt: Date,
-  encryptionKey: CryptoKey,
-  accessToken: string,
-  refreshToken: string,
-  scopes: string[],
-) {
-  return await tracer.asyncSpan('upsert-candidate-sender', async span => {
-    span.setAttributes({
-      'database.user.id': userId,
-      'database.candidate_sender.expired_at': expiredAt.toISOString(),
-    });
-    const [encryptedAccessToken, encryptedRefreshToken] = await Promise.all([
-      encryptSecret(encryptionKey, accessToken),
-      encryptSecret(encryptionKey, refreshToken),
-    ]);
-    const { rowCount } = await db
-      .insert(schema.candidateSender)
-      .values({
-        userId,
-        expiredAt,
-        accessTokenIv: Buffer.from(encryptedAccessToken.iv),
-        accessTokenCipher: Buffer.from(encryptedAccessToken.cipher),
-        refreshTokenIv: Buffer.from(encryptedRefreshToken.iv),
-        refreshTokenCipher: Buffer.from(encryptedRefreshToken.cipher),
-        scopes,
-      })
-      .onConflictDoUpdate({
-        target: schema.candidateSender.userId,
-        set: {
-          updatedAt: sql`now()`,
-          expiredAt: sql`excluded.${sql.raw(schema.candidateSender.expiredAt.name)}`,
-          accessTokenIv: sql`excluded.${sql.raw(schema.candidateSender.accessTokenIv.name)}`,
-          accessTokenCipher: sql`excluded.${sql.raw(schema.candidateSender.accessTokenCipher.name)}`,
-          refreshTokenIv: sql`excluded.${sql.raw(schema.candidateSender.refreshTokenIv.name)}`,
-          refreshTokenCipher: sql`excluded.${sql.raw(schema.candidateSender.refreshTokenCipher.name)}`,
-          scopes: sql`excluded.${sql.raw(schema.candidateSender.scopes.name)}`,
-        },
-      });
-    switch (rowCount) {
-      case 1:
-        return;
-      default:
-        fail(`upsertCandidateSender => unexpected insertion count ${rowCount}`);
-    }
-  });
-}
-
-export async function deleteCandidateSender(db: DbConnection, userId: string) {
-  return await tracer.asyncSpan('delete-candidate-sender', async span => {
-    span.setAttribute('database.user.id', userId);
-    const { rowCount } = await db
-      .delete(schema.candidateSender)
-      .where(eq(schema.candidateSender.userId, userId));
-    switch (rowCount) {
-      case 0:
-        return false;
-      case 1:
-        return true;
-      default:
-        fail(`deleteCandidateSender => unexpected delete count ${count}`);
-    }
-  });
-}
-
-export async function deleteDesignatedSender(db: DbConnection, userId: string) {
-  return await tracer.asyncSpan('delete-designated-sender', async span => {
-    span.setAttribute('database.user.id', userId);
-    const { rowCount } = await db
-      .delete(schema.designatedSender)
-      .where(eq(schema.designatedSender.candidateSenderUserId, userId));
-    switch (rowCount) {
-      case 0:
-        return false;
-      case 1:
-        return true;
-      default:
-        fail(`deleteDesignatedSender => unexpected delete count ${rowCount}`);
-    }
-  });
-}
-
-export async function upsertDesignatedSender(db: DbConnection, userId: string) {
-  return await tracer.asyncSpan('upsert-designated-sender', async span => {
-    span.setAttribute('database.user.id', userId);
-    const { rowCount } = await db
-      .insert(schema.designatedSender)
-      .values({ candidateSenderUserId: userId })
-      .onConflictDoNothing({ target: schema.designatedSender.candidateSenderUserId });
-    switch (rowCount) {
-      case 0:
-        return false;
-      case 1:
-        return true;
-      default:
-        fail(`upsertDesignatedSender => unexpected insertion count ${rowCount}`);
-    }
-  });
-}
-
-/**
- * Upserts a faculty choice for a lab in the specified draft round.
- * Uses ON CONFLICT DO UPDATE for `faculty_choice` and deletes/reinserts `faculty_choice_user`.
- * Typically invoked from within a transaction.
- */
-export async function upsertFacultyChoice(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  round: number,
-  labId: string,
-  facultyUserId: string,
-  studentUserIds: string[],
-) {
-  return await tracer.asyncSpan('upsert-faculty-choice', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.lab.id', labId);
-    span.setAttribute('database.user.id', facultyUserId);
-
-    await db
-      .insert(schema.facultyChoice)
-      .values({ draftId, round, labId, userId: facultyUserId })
-      .onConflictDoUpdate({
-        target: [
-          schema.facultyChoice.draftId,
-          schema.facultyChoice.round,
-          schema.facultyChoice.labId,
-        ],
-        set: { userId: facultyUserId, createdAt: sql`now()` },
-      });
-
-    // clear and re-insert to handle student removals and avoid orphaned rows
-    await db
-      .delete(schema.facultyChoiceUser)
-      .where(
-        and(
-          eq(schema.facultyChoiceUser.draftId, draftId),
-          eq(schema.facultyChoiceUser.labId, labId),
-          eq(schema.facultyChoiceUser.round, round),
-        ),
-      );
-
-    if (studentUserIds.length > 0) {
-      const { rowCount: facultyChoiceUserRowCount } = await db
-        .insert(schema.facultyChoiceUser)
-        .values(
-          studentUserIds.map(
-            studentUserId =>
-              ({
-                draftId,
-                round,
-                labId,
-                facultyUserId,
-                studentUserId,
-              }) satisfies schema.NewFacultyChoiceUser,
-          ),
-        );
-      strictEqual(
-        facultyChoiceUserRowCount,
-        studentUserIds.length,
-        'upsertFacultyChoice::facultyChoiceUser => unexpected insertion count',
-      );
-    }
-  });
-}
-
-/** Typically invoked from within a transaction. */
-export async function insertLotteryChoices(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  adminUserId: string,
-  assignmentUserIdToLabPairs: (readonly [string, string])[],
-  mode: 'intervention' | 'lottery',
-) {
-  return await tracer.asyncSpan('insert-lottery-choices', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    span.setAttribute('database.user.admin_id', adminUserId);
-    span.setAttribute('database.choice.mode', mode);
-    const draft = await db.query.draft.findFirst({
-      columns: { maxRounds: true },
-      where: ({ id }, { eq }) => eq(id, draftId),
-    });
-    if (typeof draft === 'undefined') return;
-
-    const labs = await db
-      .select({ id: schema.draftLabQuota.labId })
-      .from(schema.draftLabQuota)
-      .where(eq(schema.draftLabQuota.draftId, draftId));
-    if (labs.length === 0) return;
-
-    const round = mode === 'intervention' ? draft.maxRounds + 1 : null;
-    await db
-      .insert(schema.facultyChoice)
-      .values(
-        labs.map(
-          ({ id }) =>
-            ({
-              draftId,
-              round,
-              labId: id,
-              // Non-null to assert that the lottery was initiated by the admin.
-              userId: adminUserId,
-            }) satisfies schema.NewFacultyChoice,
-        ),
-      )
-      .onConflictDoUpdate({
-        target: [
-          schema.facultyChoice.draftId,
-          schema.facultyChoice.round,
-          schema.facultyChoice.labId,
-        ],
-        set: { userId: sql`excluded.${sql.raw(schema.facultyChoice.userId.name)}` },
-      });
-
-    if (assignmentUserIdToLabPairs.length > 0) {
-      const facultyUserId = mode === 'intervention' ? adminUserId : null;
-      await db.insert(schema.facultyChoiceUser).values(
-        assignmentUserIdToLabPairs.map(
-          ([studentUserId, labId]): schema.NewFacultyChoiceUser => ({
-            facultyUserId,
-            studentUserId,
-            draftId,
-            round,
-            labId,
-          }),
-        ),
-      );
-    }
-
-    return draft;
   });
 }
 
@@ -1824,93 +366,11 @@ export async function getPendingLabCountInDraft(db: DbConnection, draftId: bigin
   });
 }
 
-export async function getFacultyChoiceRecords(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-faculty-choice-records', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const facultyUser = alias(schema.user, 'faculty_user');
-    const studentUser = alias(schema.user, 'student_user');
-    return await db
-      .select({
-        draftId: schema.facultyChoice.draftId,
-        round: schema.facultyChoice.round,
-        labId: schema.facultyChoice.labId,
-        createdAt: schema.facultyChoice.createdAt,
-        userId: schema.facultyChoice.userId,
-        userEmail: facultyUser.email,
-        studentEmail: studentUser.email,
-      })
-      .from(schema.facultyChoice)
-      .leftJoin(facultyUser, eq(schema.facultyChoice.userId, facultyUser.id))
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.facultyChoice.draftId, schema.facultyChoiceUser.draftId),
-          eq(schema.facultyChoice.labId, schema.facultyChoiceUser.labId),
-          sql`${schema.facultyChoice.round} is not distinct from ${schema.facultyChoiceUser.round}`,
-        ),
-      )
-      .leftJoin(studentUser, eq(schema.facultyChoiceUser.studentUserId, studentUser.id))
-      .where(eq(schema.facultyChoice.draftId, draftId))
-      .orderBy(
-        desc(schema.facultyChoice.createdAt),
-        sql`${schema.facultyChoice.round} DESC NULLS FIRST`,
-        asc(schema.facultyChoice.labId),
-      );
-  });
-}
-
-export async function getSystemLogsExport(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-system-logs-export', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const facultyUser = alias(schema.user, 'faculty_user');
-    const studentUser = alias(schema.user, 'student_user');
-    return await db
-      .select({
-        draftId: schema.facultyChoice.draftId,
-        round: schema.facultyChoice.round,
-        labId: schema.facultyChoice.labId,
-        createdAt: schema.facultyChoice.createdAt,
-        userId: schema.facultyChoice.userId,
-        userEmail: facultyUser.email,
-        studentEmails:
-          sql`coalesce(array_agg(${studentUser.email}) filter (where ${studentUser.email} is not null), '{}')`.mapWith(
-            vals => parse(StringArray, vals),
-          ),
-      })
-      .from(schema.facultyChoice)
-      .leftJoin(facultyUser, eq(schema.facultyChoice.userId, facultyUser.id))
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.facultyChoice.draftId, schema.facultyChoiceUser.draftId),
-          eq(schema.facultyChoice.labId, schema.facultyChoiceUser.labId),
-          sql`${schema.facultyChoice.round} is not distinct from ${schema.facultyChoiceUser.round}`,
-        ),
-      )
-      .leftJoin(studentUser, eq(schema.facultyChoiceUser.studentUserId, studentUser.id))
-      .where(eq(schema.facultyChoice.draftId, draftId))
-      .groupBy(
-        schema.facultyChoice.draftId,
-        schema.facultyChoice.round,
-        schema.facultyChoice.labId,
-        schema.facultyChoice.createdAt,
-        schema.facultyChoice.userId,
-        facultyUser.email,
-      )
-      .orderBy(
-        desc(schema.facultyChoice.createdAt),
-        sql`${schema.facultyChoice.round} desc nulls first`,
-        asc(schema.facultyChoice.labId),
-      );
-  });
-}
-
 export async function getDraftAssignmentRecords(db: DbConnection, draftId: bigint) {
   return await tracer.asyncSpan('get-draft-assignment-records', async span => {
     span.setAttribute('database.draft.id', draftId.toString());
     const choice = alias(schema.facultyChoice, 'choice');
     const student = alias(schema.user, 'student');
-
     return await db
       .select({
         id: schema.facultyChoiceUser.studentUserId,
@@ -1944,336 +404,6 @@ export async function getDraftAssignmentRecords(db: DbConnection, draftId: bigin
   });
 }
 
-export async function getDraftAssignmentCountsByAttribute(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-draft-assignment-counts-by-lab', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .select({
-        labId: schema.facultyChoiceUser.labId,
-        round: schema.facultyChoiceUser.round,
-        count: count(schema.facultyChoiceUser.studentUserId),
-      })
-      .from(schema.facultyChoiceUser)
-      .where(eq(schema.facultyChoiceUser.draftId, draftId))
-      .groupBy(({ labId, round }) => [labId, round]);
-  });
-}
-
-export async function getStudentRanksExport(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-student-ranks-export', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .select({
-        createdAt: schema.studentRank.createdAt,
-        email: schema.user.email,
-        studentNumber: schema.user.studentNumber,
-        givenName: schema.user.givenName,
-        familyName: schema.user.familyName,
-        labRanks:
-          sql`coalesce(array_agg(${schema.studentRankLab.labId} ORDER BY ${schema.studentRankLab.index}) filter (where ${isNotNull(schema.studentRankLab.labId)}), '{}')`.mapWith(
-            vals => parse(StringArray, vals),
-          ),
-      })
-      .from(schema.studentRank)
-      .innerJoin(schema.user, eq(schema.studentRank.userId, schema.user.id))
-      .leftJoin(
-        schema.studentRankLab,
-        and(
-          eq(schema.studentRank.draftId, schema.studentRankLab.draftId),
-          eq(schema.studentRank.userId, schema.studentRankLab.userId),
-        ),
-      )
-      .where(eq(schema.studentRank.draftId, draftId))
-      .groupBy(schema.user.id, schema.studentRank.createdAt)
-      .orderBy(asc(schema.studentRank.createdAt));
-  });
-}
-
-export async function getDraftResultsExport(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-draft-results-export', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const facultyUser = alias(schema.user, 'faculty_user');
-    const studentUser = alias(schema.user, 'student_user');
-    return await db
-      .select({
-        studentEmail: studentUser.email,
-        studentNumber: studentUser.studentNumber,
-        studentFamilyName: studentUser.familyName,
-        studentGivenName: studentUser.givenName,
-        facultyEmail: facultyUser.email,
-        facultyFamilyName: facultyUser.familyName,
-        facultyGivenName: facultyUser.givenName,
-        round: schema.facultyChoiceUser.round,
-        lab: schema.lab.id,
-      })
-      .from(schema.facultyChoiceUser)
-      .innerJoin(schema.lab, eq(schema.facultyChoiceUser.labId, schema.lab.id))
-      .leftJoin(facultyUser, eq(schema.facultyChoiceUser.facultyUserId, facultyUser.id))
-      .innerJoin(studentUser, eq(schema.facultyChoiceUser.studentUserId, studentUser.id))
-      .where(eq(schema.facultyChoiceUser.draftId, draftId))
-      .orderBy(asc(schema.facultyChoiceUser.round), asc(schema.lab.id), asc(studentUser.email));
-  });
-}
-
-export async function getDraftEvents(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-draft-events', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .selectDistinct({
-        createdAt: schema.facultyChoice.createdAt,
-        round: schema.facultyChoice.round,
-        labId: schema.facultyChoice.labId,
-        isSystem: isNull(schema.facultyChoice.userId).mapWith(Boolean),
-      })
-      .from(schema.facultyChoice)
-      .innerJoin(schema.draft, eq(schema.facultyChoice.draftId, schema.draft.id))
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.facultyChoice.draftId, schema.facultyChoiceUser.draftId),
-          eq(schema.facultyChoice.labId, schema.facultyChoiceUser.labId),
-          or(
-            and(isNull(schema.facultyChoice.round), isNull(schema.facultyChoiceUser.round)),
-            eq(schema.facultyChoice.round, schema.facultyChoiceUser.round),
-          ),
-        ),
-      )
-      .where(
-        and(
-          eq(schema.facultyChoice.draftId, draftId),
-          or(
-            lte(schema.facultyChoice.round, schema.draft.maxRounds),
-            isNull(schema.facultyChoice.userId),
-            isNotNull(schema.facultyChoiceUser.studentUserId),
-          ),
-        ),
-      )
-      .orderBy(({ createdAt, round, labId }) => [desc(createdAt), desc(round), asc(labId)]);
-  });
-}
-
-export async function inviteNewFacultyOrStaff(
-  db: DbConnection,
-  email: string,
-  labId: string | null,
-) {
-  return await tracer.asyncSpan('invite-new-faculty-or-staff', async span => {
-    span.setAttribute('database.user.email', email);
-    if (labId !== null) span.setAttribute('database.lab.id', labId);
-    const { rowCount } = await db
-      .insert(schema.user)
-      .values({ email, labId, isAdmin: true })
-      .onConflictDoNothing({ target: schema.user.email });
-    switch (rowCount) {
-      case 0:
-        return false;
-      case 1:
-        return true;
-      default:
-        fail(`inviteNewFacultyOrStaff => unexpected insertion count ${rowCount}`);
-    }
-  });
-}
-
-export async function deleteInvitation(db: DbConnection, id: string) {
-  return await tracer.asyncSpan('delete-invitation', async span => {
-    span.setAttribute('database.user.id', id);
-    const { rowCount } = await db
-      .delete(schema.user)
-      .where(
-        and(
-          eq(schema.user.id, id),
-          eq(schema.user.isAdmin, true),
-          isNull(schema.user.googleUserId),
-        ),
-      );
-    switch (rowCount) {
-      case 0:
-        return false;
-      case 1:
-        return true;
-      default:
-        fail(`deleteInvitation => unexpected deletion count ${rowCount}`);
-    }
-  });
-}
-
-export async function syncResultsToUsers(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('sync-results-to-users', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .update(schema.user)
-      .set({ labId: schema.facultyChoiceUser.labId })
-      .from(schema.facultyChoiceUser)
-      .where(
-        and(
-          eq(schema.facultyChoiceUser.draftId, draftId),
-          eq(schema.user.id, schema.facultyChoiceUser.studentUserId),
-        ),
-      )
-      .returning({ userId: schema.user.id, labId: schema.facultyChoiceUser.labId });
-  });
-}
-
-/** Dev-only: Updates the user's admin status and lab assignment. */
-export async function updateUserRole(
-  db: DbConnection,
-  userId: string,
-  isAdmin: boolean,
-  labId: string | null,
-) {
-  return await tracer.asyncSpan('update-user-role', async span => {
-    span.setAttributes({
-      'database.user.id': userId,
-      'database.user.is_admin': isAdmin,
-    });
-    if (labId !== null) span.setAttribute('database.lab.id', labId);
-    await db.update(schema.user).set({ isAdmin, labId }).where(eq(schema.user.id, userId));
-  });
-}
-
-export interface TestUserOptions {
-  email: string;
-  googleUserId: string;
-  givenName: string;
-  familyName: string;
-  avatarUrl?: string;
-  isAdmin: boolean;
-  labId: string | null;
-}
-
-/**
- * Test helper: Creates a user with explicit Google user ID, isAdmin, and labId.
- * Combines upsertOpenIdUser + updateUserRole for test fixtures.
- */
-export async function upsertTestUser(
-  db: DrizzleTransaction,
-  { email, googleUserId, givenName, familyName, avatarUrl = '', isAdmin, labId }: TestUserOptions,
-) {
-  const { id: userId } = await upsertOpenIdUser(
-    db,
-    email,
-    googleUserId,
-    givenName,
-    familyName,
-    avatarUrl,
-  );
-  await updateUserRole(db, userId, isAdmin, labId);
-  return { id: userId, isAdmin, labId };
-}
-
-export async function getAllowlistByDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-allowlist-by-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const student = alias(schema.user, 'student');
-    const admin = alias(schema.user, 'admin');
-    return await db
-      .select({
-        draftId: schema.draftRegistrationAllowlist.draftId,
-        studentUserId: schema.draftRegistrationAllowlist.studentUserId,
-        studentEmail: student.email,
-        adminUserId: schema.draftRegistrationAllowlist.adminUserId,
-        adminGivenName: admin.givenName,
-        adminFamilyName: admin.familyName,
-        adminEmail: admin.email,
-        createdAt: schema.draftRegistrationAllowlist.createdAt,
-        submittedAt: schema.studentRank.createdAt,
-      })
-      .from(schema.draftRegistrationAllowlist)
-      .innerJoin(admin, eq(schema.draftRegistrationAllowlist.adminUserId, admin.id))
-      .innerJoin(student, eq(schema.draftRegistrationAllowlist.studentUserId, student.id))
-      .leftJoin(
-        schema.studentRank,
-        and(
-          eq(schema.draftRegistrationAllowlist.draftId, schema.studentRank.draftId),
-          eq(schema.draftRegistrationAllowlist.studentUserId, schema.studentRank.userId),
-        ),
-      )
-      .where(eq(schema.draftRegistrationAllowlist.draftId, draftId))
-      .orderBy(desc(schema.draftRegistrationAllowlist.createdAt));
-  });
-}
-
-export async function getAllowlistCountByDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-allowlist-count-by-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const { result } = await db
-      .select({ result: count(schema.draftRegistrationAllowlist.studentUserId) })
-      .from(schema.draftRegistrationAllowlist)
-      .where(eq(schema.draftRegistrationAllowlist.draftId, draftId))
-      .then(assertSingle);
-    return result;
-  });
-}
-
-export async function addToAllowlist(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  studentUserId: string,
-  adminUserId: string,
-) {
-  return await tracer.asyncSpan('add-to-allowlist', async span => {
-    span.setAttributes({
-      'database.draft.id': draftId.toString(),
-      'database.user.student_id': studentUserId,
-      'database.user.admin_id': adminUserId,
-    });
-    const result = await db
-      .insert(schema.draftRegistrationAllowlist)
-      .values({ draftId, studentUserId, adminUserId })
-      .onConflictDoNothing({
-        target: [
-          schema.draftRegistrationAllowlist.draftId,
-          schema.draftRegistrationAllowlist.studentUserId,
-        ],
-      });
-    assert(result.rowCount !== null);
-    logger.debug('added to allowlist', { rowCount: result.rowCount });
-    return result.rowCount;
-  });
-}
-
-export async function removeFromAllowlist(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  studentUserId: string,
-) {
-  return await tracer.asyncSpan('remove-from-allowlist', async span => {
-    span.setAttributes({
-      'database.draft.id': draftId.toString(),
-      'database.user.id': studentUserId,
-    });
-    return await db
-      .delete(schema.draftRegistrationAllowlist)
-      .where(
-        and(
-          eq(schema.draftRegistrationAllowlist.draftId, draftId),
-          eq(schema.draftRegistrationAllowlist.studentUserId, studentUserId),
-        ),
-      );
-  });
-}
-
-export async function isUserInAllowlist(db: DbConnection, draftId: bigint, studentUserId: string) {
-  return await tracer.asyncSpan('is-user-in-allowlist', async span => {
-    span.setAttributes({
-      'database.draft.id': draftId.toString(),
-      'database.user.id': studentUserId,
-    });
-    const result = await db
-      .select({ studentUserId: schema.draftRegistrationAllowlist.studentUserId })
-      .from(schema.draftRegistrationAllowlist)
-      .where(
-        and(
-          eq(schema.draftRegistrationAllowlist.draftId, draftId),
-          eq(schema.draftRegistrationAllowlist.studentUserId, studentUserId),
-        ),
-      )
-      .then(assertOptional);
-    return typeof result !== 'undefined';
-  });
-}
-
 export async function getUserByEmail(db: DbConnection, email: string) {
   return await tracer.asyncSpan('get-user-by-email', async span => {
     span.setAttribute('database.user.email', email);
@@ -2283,193 +413,5 @@ export async function getUserByEmail(db: DbConnection, email: string) {
       .from(schema.user)
       .where(eq(schema.user.email, email))
       .then(assertOptional);
-  });
-}
-
-export async function isRegisteredOrAssignedInDraft(
-  db: DrizzleTransaction,
-  draftId: bigint,
-  userId: string,
-) {
-  return await tracer.asyncSpan('is-registered-or-assigned-in-draft', async span => {
-    span.setAttributes({
-      'database.draft.id': draftId.toString(),
-      'database.user.id': userId,
-    });
-
-    // Check studentRank (submitted rankings)
-    const registeredResult = await db
-      .select({ userId: schema.studentRank.userId })
-      .from(schema.studentRank)
-      .innerJoin(schema.user, eq(schema.studentRank.userId, schema.user.id))
-      .where(and(eq(schema.studentRank.draftId, draftId), eq(schema.user.id, userId)))
-      .then(assertOptional);
-
-    if (typeof registeredResult !== 'undefined') {
-      logger.debug('registered', { 'student_rank.user_id': registeredResult.userId });
-      return true;
-    }
-
-    // Check facultyChoiceUser (assigned to lab)
-    const assignedResult = await db
-      .select({ studentUserId: schema.facultyChoiceUser.studentUserId })
-      .from(schema.facultyChoiceUser)
-      .innerJoin(schema.user, eq(schema.facultyChoiceUser.studentUserId, schema.user.id))
-      .where(and(eq(schema.facultyChoiceUser.draftId, draftId), eq(schema.user.id, userId)))
-      .then(assertOptional);
-
-    if (typeof assignedResult !== 'undefined') {
-      logger.debug('assigned', { 'faculty_choice.student_user_id': assignedResult.studentUserId });
-      return true;
-    }
-
-    logger.debug('not registered or assigned');
-    return false;
-  });
-}
-
-export async function getLateRegistrantsByDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-late-registrants-by-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    return await db
-      .select({
-        id: schema.user.id,
-        email: schema.user.email,
-        givenName: schema.user.givenName,
-        familyName: schema.user.familyName,
-        avatarUrl: schema.user.avatarUrl,
-        studentNumber: schema.user.studentNumber,
-        labs: sql`coalesce(array_agg(${schema.studentRankLab.labId} order by ${schema.studentRankLab.index}) filter (where ${isNotNull(schema.studentRankLab.labId)}), '{}')`
-          .mapWith(value => parse(StringArray, value))
-          .as('labs'),
-        labId: schema.facultyChoiceUser.labId,
-      })
-      .from(schema.studentRank)
-      .innerJoin(schema.user, eq(schema.studentRank.userId, schema.user.id))
-      .innerJoin(schema.draft, eq(schema.studentRank.draftId, schema.draft.id))
-      .leftJoin(
-        schema.facultyChoiceUser,
-        and(
-          eq(schema.studentRank.draftId, schema.facultyChoiceUser.draftId),
-          eq(schema.studentRank.userId, schema.facultyChoiceUser.studentUserId),
-        ),
-      )
-      .leftJoin(
-        schema.studentRankLab,
-        and(
-          eq(schema.studentRank.draftId, schema.studentRankLab.draftId),
-          eq(schema.studentRank.userId, schema.studentRankLab.userId),
-        ),
-      )
-      .where(
-        and(
-          eq(schema.studentRank.draftId, draftId),
-          lt(schema.draft.registrationClosedAt, schema.studentRank.createdAt),
-        ),
-      )
-      .groupBy(schema.user.id, schema.facultyChoiceUser.labId)
-      .orderBy(schema.user.familyName);
-  });
-}
-
-export async function getLateRegistrantsCountByDraft(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-late-registrants-count-by-draft', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const { result } = await db
-      .select({ result: count(schema.studentRank.userId) })
-      .from(schema.studentRank)
-      .innerJoin(schema.draft, eq(schema.studentRank.draftId, schema.draft.id))
-      .where(
-        and(
-          eq(schema.studentRank.draftId, draftId),
-          lt(schema.draft.registrationClosedAt, schema.studentRank.createdAt),
-        ),
-      )
-      .then(assertSingle);
-    return result;
-  });
-}
-
-export async function getDraftRegistrationTimeline(db: DbConnection, id: bigint) {
-  return await tracer.asyncSpan('get-draft-registration-timeline', async span => {
-    span.setAttribute('database.draft.id', id.toString());
-    return await db.query.studentRank.findMany({
-      columns: { createdAt: true },
-      where: ({ draftId }, { eq }) => eq(draftId, id),
-      orderBy: ({ createdAt }) => createdAt,
-    });
-  });
-}
-
-export async function getLabDemandBordaScores(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-lab-demand-borda-scores', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const studentLabCount = db
-      .select({
-        draftId: schema.studentRankLab.draftId,
-        userId: schema.studentRankLab.userId,
-        n: count().as('n'),
-      })
-      .from(schema.studentRankLab)
-      .where(eq(schema.studentRankLab.draftId, draftId))
-      .groupBy(({ draftId, userId }) => [draftId, userId])
-      .as('student_lab_count');
-    return await db
-      .select({
-        labId: schema.studentRankLab.labId,
-        bordaScore: sum(sql`${studentLabCount.n} - ${schema.studentRankLab.index} + 1`).mapWith(
-          coerceNumber,
-        ),
-      })
-      .from(schema.studentRankLab)
-      .innerJoin(
-        studentLabCount,
-        and(
-          eq(schema.studentRankLab.draftId, studentLabCount.draftId),
-          eq(schema.studentRankLab.userId, studentLabCount.userId),
-        ),
-      )
-      .where(eq(schema.studentRankLab.draftId, draftId))
-      .groupBy(({ labId }) => labId);
-  });
-}
-
-export async function getDraftPreferenceAlignment(db: DbConnection, draftId: bigint) {
-  return await tracer.asyncSpan('get-draft-preference-alignment', async span => {
-    span.setAttribute('database.draft.id', draftId.toString());
-    const studentLabCount = db
-      .select({
-        draftId: schema.studentRankLab.draftId,
-        userId: schema.studentRankLab.userId,
-        n: count().as('n'),
-      })
-      .from(schema.studentRankLab)
-      .where(eq(schema.studentRankLab.draftId, draftId))
-      .groupBy(({ draftId, userId }) => [draftId, userId])
-      .as('student_lab_count');
-    return await db
-      .select({
-        preferenceRank: schema.studentRankLab.index,
-        totalRanked: studentLabCount.n,
-        count: count(),
-      })
-      .from(schema.facultyChoiceUser)
-      .leftJoin(
-        schema.studentRankLab,
-        and(
-          eq(schema.facultyChoiceUser.draftId, schema.studentRankLab.draftId),
-          eq(schema.facultyChoiceUser.studentUserId, schema.studentRankLab.userId),
-          eq(schema.facultyChoiceUser.labId, schema.studentRankLab.labId),
-        ),
-      )
-      .leftJoin(
-        studentLabCount,
-        and(
-          eq(schema.facultyChoiceUser.draftId, studentLabCount.draftId),
-          eq(schema.facultyChoiceUser.studentUserId, studentLabCount.userId),
-        ),
-      )
-      .where(eq(schema.facultyChoiceUser.draftId, draftId))
-      .groupBy(({ preferenceRank, totalRanked }) => [preferenceRank, totalRanked]);
   });
 }
