+++ src/routes/dashboard/(draft)/drafts/[draftId]/allowlist/+server.ts
@@ -1,8 +1,11 @@
 import * as devalue from 'devalue';
+import { alias } from 'drizzle-orm/pg-core';
+import { and, desc, eq } from 'drizzle-orm';
 import { error } from '@sveltejs/kit';
 
+import * as schema from '$lib/server/database/schema';
 import { db } from '$lib/server/database';
-import { getAllowlistByDraft } from '$lib/server/database/drizzle';
+import type { DbConnection } from '$lib/server/database/drizzle';
 import { Logger } from '$lib/server/telemetry/logger';
 import { Tracer } from '$lib/server/telemetry/tracer';
 
@@ -49,5 +52,37 @@ export async function GET({ params, locals: { session } }) {
     return new Response(devalue.stringify(allowlist), {
       headers: { 'Content-Type': 'application/json' },
     });
+  });
+}
+
+async function getAllowlistByDraft(db: DbConnection, draftId: bigint) {
+  return await tracer.asyncSpan('get-allowlist-by-draft', async span => {
+    span.setAttribute('database.draft.id', draftId.toString());
+    const student = alias(schema.user, 'student');
+    const admin = alias(schema.user, 'admin');
+    return await db
+      .select({
+        draftId: schema.draftRegistrationAllowlist.draftId,
+        studentUserId: schema.draftRegistrationAllowlist.studentUserId,
+        studentEmail: student.email,
+        adminUserId: schema.draftRegistrationAllowlist.adminUserId,
+        adminGivenName: admin.givenName,
+        adminFamilyName: admin.familyName,
+        adminEmail: admin.email,
+        createdAt: schema.draftRegistrationAllowlist.createdAt,
+        submittedAt: schema.studentRank.createdAt,
+      })
+      .from(schema.draftRegistrationAllowlist)
+      .innerJoin(admin, eq(schema.draftRegistrationAllowlist.adminUserId, admin.id))
+      .innerJoin(student, eq(schema.draftRegistrationAllowlist.studentUserId, student.id))
+      .leftJoin(
+        schema.studentRank,
+        and(
+          eq(schema.draftRegistrationAllowlist.draftId, schema.studentRank.draftId),
+          eq(schema.draftRegistrationAllowlist.studentUserId, schema.studentRank.userId),
+        ),
+      )
+      .where(eq(schema.draftRegistrationAllowlist.draftId, draftId))
+      .orderBy(desc(schema.draftRegistrationAllowlist.createdAt));
   });
 }
